# api/openapi.yaml
openapi: 3.1.0
info:
  title: Swarm Agent
  description: This service implements ACO algorithm for data                 search
    and movement
  contact:
    name: Lakeside Labs
    email: chepizhko@lakeside-labs.com
  license:
    name: MIT License
    url: https://github.com/glaciation-heu/swarm-agent/blob/main/LICENSE
  version: 1.0.1
paths:
  /api/v0/create_agent:
    post:
      summary: Receive Message
      description: 'Receive message, parse it, create Swarm Agent, make a step.

        We can use the same function to receive messages from both

        Metadata Service and other Swarm Agents.'
      operationId: receive_message_api_v0_create_agent_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Receive Message Api V0 Create Agent Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v0/pheromone:
    post:
      summary: Pheromone Pointing To Neighbor
      operationId: pheromone_pointing_to_neighbor_api_v0_pheromone_post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: Body
              description: 'The neighbor where the pheromone points needsto be specified
                in the following way:{"neighbor": <neighbor-name>}'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v0/data_movement/recommendations:
    get:
      summary: Data Movement Recommendations
      description: 'Returns the most recent data movement recommendations issued by
        this node''s

        DataMovementAgent. Recommendations are advisory: they identify which neighbor

        should receive a given keyword''s data based on pheromone gradients, but no

        data is moved autonomously. An operator or CRD controller should act on these.'
      operationId: data_movement_recommendations_api_v0_data_movement_recommendations_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DataMovementRecommendation'
                type: array
                title: Response Data Movement Recommendations Api V0 Data Movement
                  Recommendations Get
  /api/v0/data_catalog:
    get:
      summary: Data Catalog
      description: 'Returns the data placement catalog written by the hub agent at
        setup time.

        Each entry contains the dataset URI, the pod name that holds it, and a

        ready-to-use SPARQL query the experiment script should send as a forward ant.

        Useful for generating diverse multi-keyword workloads and for computing the

        oracle (shortest path from any query-origin node to the data-holding node).'
      operationId: data_catalog_api_v0_data_catalog_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DataCatalogEntry'
                type: array
                title: Response Data Catalog Api V0 Data Catalog Get
  /api/v0/admin/reset_pheromones:
    post:
      summary: Reset Pheromones
      description: 'Clears the local pheromone graph on this node''s metadata service.

        Used by experiment scripts to simulate a mid-run topology or data-locality

        shift without restarting the pod (dynamic workload experiments).'
      operationId: reset_pheromones_api_v0_admin_reset_pheromones_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Reset Pheromones Api V0 Admin Reset Pheromones Post
  /metrics:
    get:
      summary: Metrics
      description: Endpoint that serves Prometheus metrics.
      operationId: metrics_metrics_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
components:
  schemas:
    DataCatalogEntry:
      properties:
        dataset_uri:
          type: string
          title: Dataset Uri
        located_at:
          type: string
          title: Located At
        sparql_query:
          type: string
          title: Sparql Query
      type: object
      required:
      - dataset_uri
      - located_at
      - sparql_query
      title: DataCatalogEntry
    DataMovementRecommendation:
      properties:
        keyword:
          type: string
          title: Keyword
        from_node:
          type: string
          title: From Node
        to_node:
          type: string
          title: To Node
        timestamp:
          type: number
          title: Timestamp
      type: object
      required:
      - keyword
      - from_node
      - to_node
      title: DataMovementRecommendation
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Message:
      properties:
        message_type:
          type: string
          enum:
          - forward
          - backward
          title: Message Type
          default: forward
        unique_id:
          type: string
          title: Unique Id
          default: ''
        sparql_query:
          type: string
          title: Sparql Query
          default: SELECT * WHERE {?s ?p ?o} LIMIT 10
        visited_nodes:
          items:
            additionalProperties:
              type: string
            type: object
          type: array
          title: Visited Nodes
          default: []
        link_costs:
          additionalProperties:
            type: number
          type: object
          title: Link Costs
          default: {}
        time_to_live:
          type: integer
          title: Time To Live
          default: 25
        keyword:
          type: string
          title: Keyword
          default: ''
        results:
          allOf:
          - $ref: '#/components/schemas/SearchResponse'
          default:
            head:
              vars: []
            results:
              bindings: []
        time_sent:
          type: number
          title: Time Sent
          default: 0.0
        time_received:
          type: number
          title: Time Received
          default: 0.0
      type: object
      title: Message
    ResponseHead:
      properties:
        vars:
          items:
            type: string
          type: array
          title: Vars
      type: object
      required:
      - vars
      title: ResponseHead
    ResponseResults:
      properties:
        bindings:
          items:
            type: object
          type: array
          title: Bindings
      type: object
      required:
      - bindings
      title: ResponseResults
    SearchResponse:
      properties:
        head:
          $ref: '#/components/schemas/ResponseHead'
        results:
          $ref: '#/components/schemas/ResponseResults'
      type: object
      required:
      - head
      - results
      title: SearchResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError